requirements.txt
numpy~=1.26.2
ai2thor~=5.0.0
jmespath~=1.0.1
opencv-python~=4.8.1.78
cffi~=1.16.0
Pillow~=10.1.0
pip~=23.0.1
attrs~=23.1.0
cryptography~=41.0.7
Jinja2~=3.1.2
setuptools~=65.5.0
pytz~=2023.3.post1
six~=1.16.0
PyYAML~=6.0.1
MarkupSafe~=2.1.3
werkzeug~=3.0.1
click~=8.1.7
blinker~=1.7.0
itsdangerous~=2.1.2
sympy~=1.12
mpmath~=1.3.0
torch~=2.1.1
packaging~=23.2
urllib3~=2.0.7
fsspec~=2023.10.0
twisted~=23.10.0
txaio~=23.1.1
requests~=2.31.0
msgpack~=1.0.7
flask~=3.0.0
Automat~=22.10.0
certifi~=2023.11.17
incremental~=22.10.0
constantly~=23.10.4
hyperlink~=21.0.0
autobahn~=23.6.2
botocore~=1.33.4
python-dateutil~=2.8.2
influxdb~=5.3.1
networkx~=3.2.1
idna~=3.6
pycparser~=2.21